Squash Vm and VMBase
Remove defunct simpletestrunnervm
