+++ ulox/ulox.core.tests/Package/Tests/BaseOnlyByteCodeLoxEngineTests.cs
@@ -9,8 +9,8 @@ namespace ULox.Tests
         public class SimpleEngine
         {
             public IProgram Program { get; private set; } = new Program();
-            private readonly VMBase _vm = new VMBase();
-            public VMBase VM => _vm;
+            private readonly Vm _vm = new Vm();
+            public Vm VM => _vm;
 
             public string Disassembly => Program.Disassembly;
 
@@ -38,7 +38,7 @@ namespace ULox.Tests
                 VM.SetGlobal(new HashedString("print"), Value.New(Print));
             }
 
-            private NativeCallResult Print(VMBase vm, int argc)
+            private NativeCallResult Print(Vm vm, int argc)
             {
                 var str = vm.GetArg(1).ToString();
                 _logger(str);
