+++ ulox/ulox.core/Package/Runtime/Library/CoreLibrary.cs
@@ -18,20 +18,20 @@
                 (nameof(str), Value.New(str))
                                         );
 
-        public NativeCallResult print(VMBase vm, int argCount)
+        public NativeCallResult print(Vm vm, int argCount)
         {
             _printer?.Invoke(vm.GetArg(1).ToString());
             return NativeCallResult.SuccessfulExpression;
         }
 
-        public NativeCallResult str(VMBase vm, int argCount)
+        public NativeCallResult str(Vm vm, int argCount)
         {
             var v = vm.GetArg(1);
             vm.PushReturn(Value.New(v.str()));
             return NativeCallResult.SuccessfulExpression;
         }
 
-        public NativeCallResult Duplicate(VMBase vm, int argCount)
+        public NativeCallResult Duplicate(Vm vm, int argCount)
         {
             vm.PushReturn(Value.Copy(vm.GetArg(1)));
             return NativeCallResult.SuccessfulExpression;
